+++ CHANGELOG.md
@@ -42,7 +42,11 @@ is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   hashes rather than vector clocks.
 - **Removed**: `Backend.getMissingChanges()`; use `Backend.getChanges()` instead.
 - **Removed**: `Backend.getChangesForActor()` since it does not fit with a hash chaining approach.
-- **Added**: `Frontend.getDeps()` returns the latest change hashes in the current document state.
+- **Added**: `Frontend.getDeps()` returns the latest change hashes in the current document state,
+  and `Frontend.getLastLocalChange()` returns the binary encoding of the last change made by the
+  local actor.
+- **Changed**: `Backend.applyLocalChange()` now returns an array of three values: the updated
+  backend state, the patch to apply to the frontend, and the binary encoding of the change.
 - **Removed**: `Automerge.Connection`, `Automerge.DocSet`, and `Automerge.WatchableDoc` have been
   moved to a separate [automerge-connection](https://github.com/automerge/automerge-connection)
   repository.